+++ OpenBudgeteer.Core/Common/Database/MySqlDatabaseContextFactory.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 using Microsoft.Extensions.Configuration;
 
@@ -9,14 +6,28 @@ namespace OpenBudgeteer.Core.Common.Database
 {
     public class MySqlDatabaseContextFactory : IDesignTimeDbContextFactory<MySqlDatabaseContext>
     {
+        private const string MysqlConnectionString = "Server=192.168.178.30;" +
+                                                     "Port=3306;" +
+                                                     "Database=openbudgeteer-dev;" +
+                                                     "User=openbudgeteer-dev;" +
+                                                     "Password=openbudgeteer-dev";
+        /*
+         * Pass arguments to args by appending them to the tool call.
+         * i.e.: dotnet ef migrations add Test -c OpenBudgeteer.Core.Common.Database.MySqlDatabaseContext -- "Server=kitana.vdaa;Port=3307;Database=openbudgeteer-dev;User=openbudgeteer-dev;Password=openbudgeteer-dev"
+         */
         public MySqlDatabaseContext CreateDbContext(string[] args)
         {
-            var optionsBuilder = new DbContextOptionsBuilder<DatabaseContext>()
-                .UseMySql("Server=192.168.178.30;" +
-                           "Port=3306;" +
-                           "Database=openbudgeteer-dev" +
-                           "User=openbudgeteer-dev" +
-                           "Password=openbudgeteer-dev");
+            var optionsBuilder = new DbContextOptionsBuilder<DatabaseContext>();
+
+            if ((args?.Length ?? 0) == 0)
+            {
+                    optionsBuilder.UseMySql(MysqlConnectionString, ServerVersion.AutoDetect(MysqlConnectionString));
+            }
+            else
+            {
+                var connectionString = string.Join(";", args);
+                optionsBuilder.UseMySql(connectionString, ServerVersion.AutoDetect(connectionString));
+            }
 
             return new MySqlDatabaseContext(optionsBuilder.Options);
         }
@@ -31,8 +42,10 @@ namespace OpenBudgeteer.Core.Common.Database
                                    $"Password={configurationSection?["Password"]}";
             var optionsBuilder = new DbContextOptionsBuilder<DatabaseContext>()
                 .UseMySql(
-                    connectionString, 
+                    connectionString,
+                    ServerVersion.AutoDetect(connectionString),
                     b => b.MigrationsAssembly("OpenBudgeteer.Core"));
+            
             return new MySqlDatabaseContext(optionsBuilder.Options);
         }
     }